as KLib is now also on github, the folder-structure has changed. adjusted code to fit the new folder-structure where KLib resides within the lib folder of the project. added _build folder to gitignore
as requested by vobject: gui-controls now support the mouse-wheel

added -Wall -Wextra -pedantic to build-settings
fixed some compiler warnings
added TODOs for audio-samples and multiple channels

added workspace settings (like beats per minute) to load/save actions
fixed sequencer listeners, added sanity checks for add (already present) and checked propper remove on delete

created ifdef to exclude optional libfftw3 from MasterTarget

+++ KSynth/sampler/Sample.h
@@ -8,7 +8,7 @@
 #ifndef SAMPLE_H_
 #define SAMPLE_H_
 
-#include <klib/fs/File.h>
+#include <KLib/fs/File.h>
 #include "../misc/DataTypes.h"
 #include <vector>
 #include <string>
@@ -41,7 +41,7 @@ public:
 
 	/** get the sample's length in milliseconds */
 	unsigned int getLengthMilliSeconds() {
-		return pcm.size() * 1000 / this->sRate;
+		return (unsigned int) pcm.size() * 1000 / this->sRate;
 	}
 
 	/** get the amount of used memory (in bytes) */
@@ -87,6 +87,7 @@ public:
 
 	/** get PCM data for the given channel */
 	std::vector<Amplitude>* getAudioData(unsigned int channel) {
+		// TODO: implement: more than one channel
 		return &pcm;
 	}
